+++ src/Game.ts
@@ -848,7 +848,7 @@ export class Game {
 		this._mainParameter = undefined;
 		this._configuration = gameConfiguration;
 		this._assetManager = new AssetManager(this, gameConfiguration.assets, gameConfiguration.audio, gameConfiguration.moduleMainScripts);
-		this._moduleManager = new ModuleManager(this._runtimeValueBase, this._assetManager);
+		this._moduleManager = undefined!;
 
 		const operationPluginsField = gameConfiguration.operationPlugins || [];
 		this.operationPluginManager = new OperationPluginManager(this, param.operationPluginViewInfo || null, operationPluginsField);
@@ -1400,6 +1400,7 @@ export class Game {
 
 		this._isTerminated = false;
 		this.vars = {};
+		this._moduleManager = new ModuleManager(this._runtimeValueBase, this._assetManager);
 
 		this.surfaceAtlasSet.destroy();
 		this.surfaceAtlasSet = new SurfaceAtlasSet({ resourceFactory: this.resourceFactory });
@@ -1524,6 +1525,7 @@ export class Game {
 		this._postTickTasks = [];
 		this.surfaceAtlasSet.destroy();
 		this.surfaceAtlasSet = undefined!;
+		this._moduleManager = undefined!;
 	}
 
 	/**